+++ internal/api/net.go
@@ -20,21 +20,26 @@ import (
 )
 
 const (
-	// ContentTypeJSON tbd
+	// ContentTypeJSON
 	ContentTypeJSON = "application/json"
 
-	// ContentTypeGeoJSON tbd
+	// ContentTypeGeoJSON
 	ContentTypeGeoJSON = "application/geo+json"
 
-	// ContentTypeHTML tbd
+	// ContentTypeHTML
 	ContentTypeHTML = "text/html"
 
+	// ContentTypeHTML
+	ContentTypeOpenAPI = "application/vnd.oai.openapi+json;version=3.0"
+
+	// FormatJSON code and extension for JSON
 	FormatJSON = "json"
 
+	// FormatHTML code and extension for HTML
 	FormatHTML = "html"
 )
 
-// ContentType tbd
+// ContentType gets the contentType for a data format
 func ContentType(format string) string {
 	switch format {
 	case FormatJSON:
@@ -45,6 +50,7 @@ func ContentType(format string) string {
 	return ""
 }
 
+// PathFormat gets the data format for a path
 func PathFormat(url *url.URL) string {
 	path := url.EscapedPath()
 	if strings.HasSuffix(path, ".html") {
@@ -53,6 +59,7 @@ func PathFormat(url *url.URL) string {
 	return FormatJSON
 }
 
+// RequestedFormat gets the format for a request from extension or headers
 func RequestedFormat(r *http.Request) string {
 	// first check explicit path
 	path := r.URL.EscapedPath()
@@ -64,13 +71,14 @@ func RequestedFormat(r *http.Request) string {
 	}
 	// Use Accept header if present
 	hdrAccept := r.Header.Get("Accept")
-	//fmt.Println(hdrAccept)
+	//fmt.Println("Accept:" + hdrAccept)
 	if strings.Index(hdrAccept, ContentTypeHTML) >= 0 {
 		return FormatHTML
 	}
 	return FormatJSON
 }
 
+// PathStripFormat removes a format extension from a path
 func PathStripFormat(path string) string {
 	if strings.HasSuffix(path, ".html") || strings.HasSuffix(path, ".json") {
 		return path[0 : len(path)-5]
@@ -78,6 +86,7 @@ func PathStripFormat(path string) string {
 	return path
 }
 
+// URLQuery gets the query part of a URL
 func URLQuery(url *url.URL) string {
 	uri := url.RequestURI()
 	qloc := strings.Index(uri, "?")