+++ internal/data/catalog.go
@@ -27,6 +27,9 @@ type Catalog interface {
 	// It returns nil if the table does not exist
 	TableByName(name string) (*Table, error)
 
+	// TableReload reloads volatile table data
+	TableReload(name string)
+
 	// TableFeatures returns an array of the JSON for the features in a table
 	// It returns nil if the table does not exist
 	TableFeatures(ctx context.Context, name string, param *QueryParam) ([]string, error)
@@ -72,6 +75,7 @@ type QueryParam struct {
 	Filter []*FilterCond
 	// Columns is the list of columns to return
 	Columns       []string
+	GroupBy       []string
 	OrderBy       []Ordering
 	Precision     int
 	TransformFuns []TransformFunction